backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.0
psycopg2-binary==2.9.9
alembic==1.13.0
pgvector==0.2.5

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.8.0
pydantic-settings==2.3.0
email-validator==2.2.0

# AI/ML
openai==1.35.0
langchain==0.2.0
langchain-openai==0.1.0
tiktoken==0.7.0

# Cache/Queue
redis==5.0.0
celery==5.4.0

# Utils
python-dotenv==1.0.0
httpx==0.27.0
